Convolve HRF after div norm in DN model
