+++ src/prfmodel/models/prf/_div_norm.py
@@ -1,237 +1,15 @@
 """Divisive normalization population receptive field models."""
 
-from typing import cast
 import pandas as pd
-from keras import ops
-from prfmodel._docstring import doc
 from prfmodel.impulse import DerivativeTwoGammaImpulse
-from prfmodel.impulse import convolve_prf_impulse_response
 from prfmodel.impulse.base import BaseImpulse
-from prfmodel.models.base import BaseCanonical
-from prfmodel.models.base import BasePopulationResponse
 from prfmodel.models.base import BaseStimulusEncoder
 from prfmodel.regressors.base import BaseRegressors
-from prfmodel.regressors.base import _normalize_regressors_model
-from prfmodel.regressors.base import _validate_regressors_argument
-from prfmodel.scaling import DivNormAmplitude
+from prfmodel.scaling import Baseline
 from prfmodel.scaling.base import BaseScaling
-from prfmodel.stimuli._prf import PRFStimulus
-from prfmodel.typing import Tensor
-from prfmodel.utils import get_dtype
 from ._gaussian import Gaussian2DPRFResponse
 from ._stimulus_encoding import PRFStimulusEncoder
-
-
-class DivNormPRFModel(BaseCanonical[PRFStimulus]):
-    r"""
-    Divisive normalization population receptive field model.
-
-    Receives two independent pRF responses (activation and normalization) and combines them
-    via the divisive normalization formula.
-    Parameters that should be shared between both responses (e.g. the pRF centre
-    ``mu_x``, ``mu_y``) are listed in ``shared_params`` and appear once in
-    :attr:`parameter_names` without a suffix. All remaining pRF parameters are suffixed
-    with ``_activation`` or ``_normalization`` respectively.
-
-    Parameters
-    ----------
-    activation_prf_model : BaseResponse
-        pRF response model (activation).
-    normalization_prf_model : BaseResponse
-        pRF response model (normalization).
-    shared_params : list of str, default=["mu_x", "mu_y"]
-        Names of pRF parameters that are shared between the two responses.  Each name
-        must appear in *both* ``activation_prf_model.parameter_names`` and
-        ``normalization_prf_model.parameter_names``.
-    %(model_encoding_prf)s
-    %(model_impulse)s
-    scaling_model : BaseScaling or type or None, default=DivNormAmplitude, optional
-        A scaling model class or instance. Model classes will be instantiated during initialization.
-        The default creates a :class:`~prfmodel.scaling.DivNormAmplitude` instance.
-    %(model_regressors)s
-
-    Notes
-    -----
-    The predicted response is:
-
-    .. math::
-
-        p_{\text{DN}} = \frac{(a R_1 \cdot S + b)}{(c R_2 \cdot S + d)} - \frac{b}{d}
-
-    Where `R_1` and `R_2` are the activation and normalization pRF responses, `S` is the stimulus.
-    The :math:`-b/d` term ensures a zero response in the absence of a stimulus.
-
-    """
-
-    def __init__(  # noqa: PLR0913
-        self,
-        activation_prf_model: BasePopulationResponse,
-        normalization_prf_model: BasePopulationResponse,
-        shared_params: list[str] | None = None,
-        encoding_model: BaseStimulusEncoder | type[BaseStimulusEncoder] = PRFStimulusEncoder,
-        impulse_model: BaseImpulse | type[BaseImpulse] | None = DerivativeTwoGammaImpulse,
-        scaling_model: BaseScaling | type[BaseScaling] | None = DivNormAmplitude,
-        regressors_model: BaseRegressors | list[BaseRegressors] | None = None,
-    ):
-        if shared_params is None:
-            shared_params = ["mu_x", "mu_y"]
-
-        if encoding_model is not None and isinstance(encoding_model, type):
-            encoding_model = encoding_model()
-
-        if impulse_model is not None and isinstance(impulse_model, type):
-            impulse_model = impulse_model()
-
-        if scaling_model is not None and isinstance(scaling_model, type):
-            scaling_model = scaling_model()
-
-        regressors_model = _normalize_regressors_model(regressors_model)
-
-        act_names = activation_prf_model.parameter_names
-        norm_names = normalization_prf_model.parameter_names
-        invalid = [p for p in shared_params if p not in act_names or p not in norm_names]
-        if invalid:
-            msg = (
-                f"DivNormPRFModel: shared_params {invalid} not found in both "
-                f"activation_prf_model.parameter_names {act_names} and "
-                f"normalization_prf_model.parameter_names {norm_names}"
-            )
-            raise ValueError(msg)
-
-        self._shared_params = shared_params
-
-        super().__init__(
-            activation_prf_model=activation_prf_model,
-            normalization_prf_model=normalization_prf_model,
-            encoding_model=encoding_model,
-            impulse_model=impulse_model,
-            scaling_model=scaling_model,
-            regressors_model=regressors_model,
-        )
-
-    @property
-    def parameter_names(self) -> list[str]:
-        """
-        Names of parameters used by the model.
-
-        Shared parameters appear once (no suffix). Response-specific parameters are suffixed
-        with ``_activation`` or ``_normalization``.
-
-        """
-        shared = set(self._shared_params)
-        act_model = cast("BasePopulationResponse", self.models["activation_prf_model"])
-        norm_model = cast("BasePopulationResponse", self.models["normalization_prf_model"])
-
-        param_names: list[str] = []
-
-        # Activation model params: shared appear as-is, non-shared get _activation suffix
-        for p in act_model.parameter_names:
-            if p in shared:
-                param_names.append(p)
-            else:
-                param_names.append(f"{p}_activation")
-
-        # Normalization model non-shared params get _normalization suffix
-        param_names.extend(f"{p}_normalization" for p in norm_model.parameter_names if p not in shared)
-
-        # Encoding, impulse, and scaling model params
-        for key, model in self.models.items():
-            if key not in ("activation_prf_model", "normalization_prf_model") and model is not None:
-                param_names.extend(model.parameter_names)
-
-        return list(dict.fromkeys(param_names))
-
-    def _predict_single_response(
-        self,
-        stimulus: PRFStimulus,
-        parameters: pd.DataFrame,
-        suffix: str,
-        dtype: str,
-    ) -> Tensor:
-        """Run one pRF response (pRF response -> encoding -> optional impulse convolution)."""
-        prf_model = cast("BasePopulationResponse", self.models[f"{suffix}_prf_model"])
-        shared = set(self._shared_params)
-
-        # Build a parameter slice for this pRF model: copy all params, then
-        # overwrite non-shared params from the suffixed columns.
-        params_single = parameters.copy()
-        for p in prf_model.parameter_names:
-            if p not in shared:
-                params_single[p] = parameters[f"{p}_{suffix}"]
-
-        response = prf_model(stimulus, params_single, dtype=dtype)
-        encoding_model = cast("BaseStimulusEncoder", self.models["encoding_model"])
-        response = encoding_model(stimulus, response, parameters, dtype=dtype)
-
-        if self.models["impulse_model"] is not None:
-            impulse_model = cast("BaseImpulse", self.models["impulse_model"])
-            impulse_response = impulse_model(parameters, dtype=dtype)
-            response = convolve_prf_impulse_response(response, impulse_response, dtype=dtype)
-
-        return response
-
-    def predict_responses(
-        self,
-        stimulus: PRFStimulus,
-        parameters: pd.DataFrame,
-        dtype: str | None = None,
-    ) -> Tensor:
-        """
-        Predict each of the two responses before applying the DN formula.
-
-        Returns
-        -------
-        Tensor
-            Stacked predictions of shape (num_voxels, 2, num_frames), where axis 1 index 0
-            is the activation response and index 1 is the normalization response.
-
-        """
-        dtype = get_dtype(dtype)
-        p1 = self._predict_single_response(stimulus, parameters, "activation", dtype)
-        p2 = self._predict_single_response(stimulus, parameters, "normalization", dtype)
-        return ops.stack([p1, p2], axis=1)
-
-    @doc
-    def __call__(
-        self,
-        stimulus: PRFStimulus,
-        parameters: pd.DataFrame,
-        regressors: pd.DataFrame | None = None,
-        dtype: str | None = None,
-    ) -> Tensor:
-        """
-        Predict the DN model response to a stimulus.
-
-        Parameters
-        ----------
-        %(stimulus_prf)s
-        %(parameters)s
-        %(regressors_canonical)s
-        %(dtype)s
-
-        Returns
-        -------
-        %(predicted_response_2d)s
-
-        """
-        dtype = get_dtype(dtype)
-        regressors_model = self.models["regressors_model"]
-        _validate_regressors_argument(regressors_model, regressors)
-
-        stacked = self.predict_responses(stimulus, parameters, dtype=dtype)
-
-        if self.models["scaling_model"] is not None:
-            scaling_model = cast("BaseScaling", self.models["scaling_model"])
-            response = scaling_model(stacked, parameters, dtype=dtype)
-        else:
-            # TODO: Is this a sensible default? Or what would word best in the absence of a scaling model?
-            response = stacked[:, 0] / stacked[:, 1]
-
-        if regressors_model is not None and regressors is not None:
-            regressors_model = cast("BaseRegressors", regressors_model)
-            response = response + regressors_model(regressors, parameters, dtype=dtype)
-
-        return response
+from .canonical import DivNormPRFModel
 
 
 class DivNormGaussian2DPRFModel(DivNormPRFModel):
@@ -249,9 +27,9 @@ class DivNormGaussian2DPRFModel(DivNormPRFModel):
     ----------
     %(model_encoding_prf)s
     %(model_impulse)s
-    scaling_model : BaseScaling or type or None, default=DivNormAmplitude, optional
+    scaling_model : BaseScaling or type or None, default=Baseline, optional
         A scaling model class or instance. Model classes will be instantiated during initialization.
-        The default creates a :class:`~prfmodel.scaling.DivNormAmplitude` instance.
+        The default creates a :class:`~prfmodel.scaling.Baseline` instance.
     %(model_regressors)s
 
     Notes
@@ -325,7 +103,7 @@ class DivNormGaussian2DPRFModel(DivNormPRFModel):
         self,
         encoding_model: BaseStimulusEncoder | type[BaseStimulusEncoder] = PRFStimulusEncoder,
         impulse_model: BaseImpulse | type[BaseImpulse] | None = DerivativeTwoGammaImpulse,
-        scaling_model: BaseScaling | type[BaseScaling] | None = DivNormAmplitude,
+        scaling_model: BaseScaling | type[BaseScaling] | None = Baseline,
         regressors_model: BaseRegressors | list[BaseRegressors] | None = None,
     ):
         super().__init__(
