+++ tests/models/prf/test_div_norm/test_predict_regression_None_None_.csv
@@ -1,0 +1,26 @@
+,response_0,response_1,response_2
+0,0.10359404981136322,0.27885603904724121,-0.026552103459835052
+1,0.15457205474376678,0.25968906283378601,-0.01368357241153717
+2,0.12603987753391266,0.16189886629581451,-0.020860046148300171
+3,0.053959950804710388,0.064250819385051727,-0.040795765817165375
+4,0.0088698584586381912,0.012599889189004898,-0.062913715839385986
+5,0.00040654648910276592,-0.0017411569133400917,-0.07854500412940979
+6,3.9192404983623419e-06,-0.0025323387235403061,-0.084997020661830902
+7,7.5876451788303712e-09,-0.0012809066101908684,-0.083482563495635986
+8,3.2887887291971341e-12,-0.00046424660831689835,-0.074739903211593628
+9,6.3830993389320283e-17,-9.3131326138973236e-05,-0.049149826169013977
+10,1.3518144634582086e-21,-1.0838732123374939e-05,-0.01411612331867218
+11,0,0,0
+12,0,0,0
+13,0,0,0
+14,1.7273659360387194e-15,-0.00010810606181621552,-0.032184936106204987
+15,1.3219764692651292e-11,-0.00050940457731485367,-0.062807723879814148
+16,2.3913198532454771e-08,-0.0014463020488619804,-0.078571140766143799
+17,2.2093789084465243e-05,-0.002769865095615387,-0.085133165121078491
+18,0.0013360134325921535,0.0005527883768081665,-0.082875706255435944
+19,0.018361330032348633,0.024488111957907677,-0.071850955486297607
+20,0.078394033014774323,0.092752620577812195,-0.052249740809202194
+21,0.14290860295295715,0.19839459657669067,-0.029837578535079956
+22,0.14870943129062653,0.27897220849990845,-0.015174910426139832
+23,0.091122768819332123,0.27061358094215393,-0.016639061272144318
+24,0.025146018713712692,0.18034231662750244,-0.033341355621814728
