+++ tests/models/prf/test_div_norm/test_predict_regression_None_temporal_model1_.csv
@@ -1,0 +1,26 @@
+,response_0,response_1,response_2
+0,-0.39640593528747559,0.77885603904724121,1.9734479188919067
+1,-0.34542793035507202,0.7596890926361084,1.986316442489624
+2,-0.37396013736724854,0.66189885139465332,1.9791399240493774
+3,-0.44604003429412842,0.56425082683563232,1.9592041969299316
+4,-0.49113014340400696,0.5125998854637146,1.9370863437652588
+5,-0.49959346652030945,0.49825882911682129,1.9214550256729126
+6,-0.49999606609344482,0.49746766686439514,1.9150029420852661
+7,-0.5,0.49871909618377686,1.9165174961090088
+8,-0.5,0.49953573942184448,1.925260066986084
+9,-0.5,0.49990686774253845,1.9508501291275024
+10,-0.5,0.49998915195465088,1.9858838319778442
+11,-0.5,0.5,2
+12,-0.5,0.5,2
+13,-0.5,0.5,2
+14,-0.5,0.49989190697669983,1.9678150415420532
+15,-0.5,0.49949058890342712,1.9371923208236694
+16,-0.49999997019767761,0.49855369329452515,1.921428918838501
+17,-0.49997791647911072,0.49723014235496521,1.9148668050765991
+18,-0.49866399168968201,0.50055277347564697,1.9171242713928223
+19,-0.48163866996765137,0.52448809146881104,1.9281489849090576
+20,-0.42160597443580627,0.59275263547897339,1.9477502107620239
+21,-0.35709139704704285,0.69839459657669067,1.9701623916625977
+22,-0.35129058361053467,0.77897220849990845,1.9848251342773438
+23,-0.40887722373008728,0.77061355113983154,1.9833608865737915
+24,-0.4748539924621582,0.68034231662750244,1.9666585922241211
