+++ tests/models/prf/test_div_norm/test_predict_regression_impulse_model1_None_.csv
@@ -1,0 +1,26 @@
+,response_0,response_1,response_2
+0,0.10359404236078262,0.27885600924491882,-0.026552099734544754
+1,0.10359404236078262,0.27885600924491882,-0.026552099734544754
+2,0.10379479825496674,0.27878051996231079,-0.026501420885324478
+3,0.10603179037570953,0.27751210331916809,-0.025936627760529518
+4,0.11146631091833115,0.26997637748718262,-0.024570433422923088
+5,0.11597806960344315,0.24728879332542419,-0.023550411686301231
+6,0.11200475692749023,0.20290741324424744,-0.025335339829325676
+7,0.095004133880138397,0.13834486901760101,-0.032155327498912811
+8,0.066611193120479584,0.062896750867366791,-0.044484015554189682
+9,0.032637882977724075,-0.011166462674736977,-0.06060304120182991
+10,-0.00047205938608385623,-0.073502905666828156,-0.077179960906505585
+11,-0.027951525524258614,-0.11817004531621933,-0.089928604662418365
+12,-0.047381151467561722,-0.14372672140598297,-0.094374425709247589
+13,-0.058378398418426514,-0.15207169950008392,-0.087810657918453217
+14,-0.061942748725414276,-0.14697851240634918,-0.071106255054473877
+15,-0.059800628572702408,-0.13283659517765045,-0.048288088291883469
+16,-0.053881194442510605,-0.1137724369764328,-0.025517897680401802
+17,-0.045956384390592575,-0.093158110976219177,-0.0095799006521701813
+18,-0.037439238280057907,-0.073436245322227478,-0.0051323045045137405
+19,-0.02930828370153904,-0.0561324842274189,-0.012843739241361618
+20,-0.022025445476174355,-0.041741698980331421,-0.02963620238006115
+21,-0.014891977421939373,-0.028962863609194756,-0.050181850790977478
+22,-0.0045317932963371277,-0.013195795007050037,-0.068557091057300568
+23,0.014666050672531128,0.013826259411871433,-0.079701103270053864
+24,0.045585963875055313,0.060101918876171112,-0.080790206789970398
