+++ tests/models/prf/test_div_norm/test_predict_regression_impulse_model1_temporal_model1_.csv
@@ -1,0 +1,26 @@
+,response_0,response_1,response_2
+0,-0.39640596508979797,0.77885603904724121,1.9734479188919067
+1,-0.39640596508979797,0.77885603904724121,1.9734479188919067
+2,-0.39620518684387207,0.77878051996231079,1.9734985828399658
+3,-0.39396822452545166,0.7775120735168457,1.9740633964538574
+4,-0.38853368163108826,0.76997637748718262,1.9754295349121094
+5,-0.38402193784713745,0.74728882312774658,1.9764496088027954
+6,-0.38799524307250977,0.70290744304656982,1.9746646881103516
+7,-0.40499585866928101,0.63834488391876221,1.9678447246551514
+8,-0.43338879942893982,0.562896728515625,1.95551598072052
+9,-0.46736210584640503,0.48883354663848877,1.9393969774246216
+10,-0.50047206878662109,0.42649710178375244,1.9228200912475586
+11,-0.52795153856277466,0.38182994723320007,1.9100713729858398
+12,-0.54738116264343262,0.35627329349517822,1.905625581741333
+13,-0.55837839841842651,0.34792828559875488,1.9121893644332886
+14,-0.56194275617599487,0.35302150249481201,1.9288938045501709
+15,-0.55980062484741211,0.36716341972351074,1.951711893081665
+16,-0.55388116836547852,0.38622754812240601,1.9744820594787598
+17,-0.54595637321472168,0.40684187412261963,1.9904201030731201
+18,-0.53743922710418701,0.42656373977661133,1.9948676824569702
+19,-0.5293082594871521,0.44386750459671021,1.9871562719345093
+20,-0.522025465965271,0.45825830101966858,1.9703638553619385
+21,-0.51489198207855225,0.47103714942932129,1.9498181343078613
+22,-0.50453180074691772,0.48680421710014343,1.9314428567886353
+23,-0.48533394932746887,0.51382625102996826,1.9202989339828491
+24,-0.45441403985023499,0.56010192632675171,1.9192098379135132
