--- tests/models/prf/test_divisive_normalization/test_predict_regression_None_None_.csv
@@ -1,51 +1,0 @@
-,response_0,response_1,response_2
-0,,,
-1,,,
-2,,,
-3,,,
-4,,,
-5,0.13756550848484039,1.4875837564468384,1.402523398399353
-6,0.19682765007019043,1.6194617748260498,1.5162602663040161
-7,0.29467174410820007,1.7643049955368042,1.6491246223449707
-8,0.42780932784080505,1.8940478563308716,1.7804235219955444
-9,0.65331476926803589,2.0855405330657959,1.989840030670166
-10,0.95678317546844482,2.2597036361694336,2.2207639217376709
-11,1.2951804399490356,2.3550868034362793,2.4222538471221924
-12,1.6228553056716919,2.3655796051025391,2.5861608982086182
-13,1.8832283020019531,2.2865760326385498,2.7003858089447021
-14,2.0232625007629395,2.1284050941467285,2.7574295997619629
-15,2.015012264251709,1.9075782299041748,2.7541208267211914
-16,1.8592445850372314,1.6448262929916382,2.6903805732727051
-17,1.5885515213012695,1.3660151958465576,2.5699193477630615
-18,1.25639808177948,1.0914689302444458,2.4014661312103271
-19,0.92177855968475342,0.84168881177902222,2.1963481903076172
-20,0.62309616804122925,0.62325263023376465,1.963674783706665
-21,0.41196209192276001,0.4675152599811554,1.7660430669784546
-22,0.28019285202026367,0.36963072419166565,1.6319596767425537
-23,0.1897282749414444,0.29421406984329224,1.5045444965362549
-24,0.13252107799053192,0.23913940787315369,1.3910173177719116
-25,,,
-26,,,
-27,,,
-28,0.099187187850475311,0.22035032510757446,0.7716338038444519
-29,0.13939094543457031,0.26711398363113403,0.85571974515914917
-30,0.210654616355896,0.3380451500415802,0.96848922967910767
-31,0.32074633240699768,0.43223413825035095,1.0994207859039307
-32,0.48913455009460449,0.56671774387359619,1.271215558052063
-33,0.76467972993850708,0.78448677062988281,1.5354589223861694
-34,1.1041585206985474,1.0467256307601929,1.8169350624084473
-35,1.4797496795654297,1.3485649824142456,2.1065425872802734
-36,1.8295310735702515,1.6689229011535645,2.3853976726531982
-37,2.0941412448883057,1.9874503612518311,2.6421089172363281
-38,2.2188689708709717,2.2767927646636963,2.8599307537078857
-39,2.1824543476104736,2.5146744251251221,3.0316839218139648
-40,1.9869413375854492,2.6731557846069336,3.1414144039154053
-41,1.6768286228179932,2.7358031272888184,3.1848993301391602
-42,1.3090523481369019,2.6950955390930176,3.1564524173736572
-43,0.94714647531509399,2.5576827526092529,3.0618717670440674
-44,0.63093119859695435,2.3357532024383545,2.903073787689209
-45,0.3893391489982605,2.0542054176330566,2.693443775177002
-46,0.22079181671142578,1.7411863803863525,2.4452164173126221
-47,0.11525400727987289,1.4177378416061401,2.168891429901123
-48,0.055441949516534805,1.1133811473846436,1.8833632469177246
-49,0.024426396936178207,0.84011495113372803,1.5982884168624878
