--- tests/models/prf/test_divisive_normalization/test_predict_regression_None_temporal_model1_.csv
@@ -1,51 +1,0 @@
-,response_0,response_1,response_2
-0,0,0,0
-1,0,0,0
-2,0,0,0
-3,0,0,0
-4,0,0,0
-5,0.11648540943861008,1.1205953359603882,0.54172408580780029
-6,0.19702234864234924,1.4004189968109131,0.75476264953613281
-7,0.30876252055168152,1.5903241634368896,0.91154932975769043
-8,0.45479795336723328,1.7332640886306763,1.0399757623672485
-9,0.7006300687789917,1.9293152093887329,1.2297803163528442
-10,1.0279113054275513,2.0957939624786377,1.4253103733062744
-11,1.3972769975662231,2.1941947937011719,1.6059305667877197
-12,1.7512696981430054,2.199967622756958,1.742927074432373
-13,2.0369405746459961,2.1267886161804199,1.8480441570281982
-14,2.1868834495544434,1.9676030874252319,1.8920605182647705
-15,2.1804463863372803,1.754953145980835,1.8949196338653564
-16,2.0083169937133789,1.4940496683120728,1.8334870338439941
-17,1.7163023948669434,1.2247079610824585,1.7345168590545654
-18,1.3530738353729248,0.95389354228973389,1.5823664665222168
-19,0.99162006378173828,0.7129366397857666,1.4090516567230225
-20,0.66662365198135376,0.49915251135826111,1.202720046043396
-21,0.43843460083007812,0.34838607907295227,1.0301384925842285
-22,0.29170864820480347,0.2475757896900177,0.88908755779266357
-23,0.18949007987976074,0.16840538382530212,0.74394655227661133
-24,0.10335864871740341,0.087204091250896454,0.48384279012680054
-25,0,0,0
-26,0,0,0
-27,0,0,0
-28,0.045817095786333084,0.036996878683567047,0.055584847927093506
-29,0.13150429725646973,0.13037535548210144,0.20453810691833496
-30,0.21420580148696899,0.20785430073738098,0.31750178337097168
-31,0.33739548921585083,0.30626288056373596,0.44389611482620239
-32,0.51975339651107788,0.43686553835868835,0.59318697452545166
-33,0.81871986389160156,0.64807420969009399,0.82101678848266602
-34,1.1843000650405884,0.89777696132659912,1.0557807683944702
-35,1.5939177274703979,1.1933038234710693,1.3088773488998413
-36,1.9711312055587769,1.4997845888137817,1.5414342880249023
-37,2.2616455554962158,1.8142622709274292,1.7691121101379395
-38,2.3944034576416016,2.0909492969512939,1.9493556022644043
-39,2.3579013347625732,2.3287067413330078,2.1056272983551025
-40,2.1421926021575928,2.4768421649932861,2.1912858486175537
-41,1.8080981969833374,2.5445992946624756,2.2379324436187744
-42,1.4064027070999146,2.4982640743255615,2.2042844295501709
-43,1.0162831544876099,2.3705799579620361,2.1315422058105469
-44,0.67273849248886108,2.1480991840362549,1.9861044883728027
-45,0.41364020109176636,1.8796399831771851,1.8138556480407715
-46,0.23188099265098572,1.5697640180587769,1.5924890041351318
-47,0.12000987678766251,1.2603522539138794,1.3622728586196899
-48,0.056484811007976532,0.96211075782775879,1.1124155521392822
-49,0.024408278986811638,0.70161354541778564,0.8744354248046875
