--- tests/models/prf/test_divisive_normalization/test_predict_regression_impulse_model1_None_.csv
@@ -1,51 +1,0 @@
-,response_0,response_1,response_2
-0,,,
-1,,,
-2,,,
-3,,,
-4,,,
-5,0.13756550848484039,1.4875837564468384,1.4025232791900635
-6,0.13756550848484039,1.4875838756561279,1.4025232791900635
-7,0.29349395632743835,1.6709431409835815,1.6021782159805298
-8,0.11682251840829849,1.4524905681610107,1.3740979433059692
-9,0.16394782066345215,1.5451748371124268,1.4530876874923706
-10,0.21086426079273224,1.6240103244781494,1.5210907459259033
-11,0.27021187543869019,1.7034205198287964,1.5917539596557617
-12,0.34990039467811584,1.7917941808700562,1.6730690002441406
-13,0.45495763421058655,1.889893651008606,1.7675304412841797
-14,0.59041696786880493,1.9964118003845215,1.8777508735656738
-15,0.75472575426101685,2.1037907600402832,2.0027363300323486
-16,0.94266271591186523,2.1995043754577637,2.1403789520263672
-17,1.141670823097229,2.2662501335144043,2.2854864597320557
-18,1.3358259201049805,2.283743143081665,2.4305999279022217
-19,1.5070134401321411,2.2335164546966553,2.5650317668914795
-20,1.6391533613204956,2.1018900871276855,2.6754772663116455
-21,1.7208535671234131,1.8855483531951904,2.7468974590301514
-22,1.7463490962982178,1.5907386541366577,2.7635340690612793
-23,1.7159333229064941,1.2336121797561646,2.7119688987731934
-24,1.6348420381546021,0.83693778514862061,2.5861940383911133
-25,1.5125553607940674,0.42223814129829407,2.3945982456207275
-26,1.3571510314941406,-0.018733497709035873,2.1496427059173584
-27,1.1654967069625854,-0.58067822456359863,1.8424117565155029
-28,0.89896804094314575,-1.7158265113830566,1.3801419734954834
-29,0.36827141046524048,-11.497485160827637,0.16082420945167542
-30,-2.8545641899108887,4.0898594856262207,15.565199851989746
-31,3.6621627807617188,1.9958240985870361,3.7056891918182373
-32,2.2929725646972656,1.4432083368301392,2.9497964382171631
-33,1.9914580583572388,1.2784576416015625,2.7343387603759766
-34,1.9309375286102295,1.5645045042037964,2.7748613357543945
-35,2.0304863452911377,-1.4939780235290527,3.1996352672576904
-36,2.3592448234558105,0.29454338550567627,5.8562116622924805
-37,3.4098701477050781,0.5448262095451355,-1.7781995534896851
-38,-101.91951751708984,0.72653788328170776,0.70435327291488647
-39,-0.2664000391960144,0.92635262012481689,1.3090677261352539
-40,0.82766562700271606,1.1654986143112183,1.669636607170105
-41,1.2998000383377075,1.4440023899078369,1.9688283205032349
-42,1.6022354364395142,1.7516424655914307,2.2522995471954346
-43,1.8119930028915405,2.0666604042053223,2.5291454792022705
-44,1.9472925662994385,2.3624513149261475,2.7947165966033936
-45,2.0127973556518555,2.6097562313079834,3.0358695983886719
-46,2.011723518371582,2.7817893028259277,3.2353148460388184
-47,1.9493657350540161,2.8578433990478516,3.3744981288909912
-48,1.8316726684570312,2.8245232105255127,3.4359781742095947
-49,1.6638295650482178,2.6792657375335693,3.4069373607635498
