--- tests/models/prf/test_divisive_normalization/test_predict_regression_impulse_model1_temporal_model1_.csv
@@ -1,51 +1,0 @@
-,response_0,response_1,response_2
-0,0,0,0
-1,0,0,0
-2,0,0,0
-3,0,0,0
-4,0,0,0
-5,-3.4742361209051431e-20,0,0
-6,-0.00078688072971999645,-0.0097997188568115234,-0.002907872200012207
-7,-0.0019306296017020941,-0.022527322173118591,-0.0057379007339477539
-8,0.016910405829548836,0.24850067496299744,0.077282369136810303
-9,0.085307665169239044,0.83680737018585205,0.33265388011932373
-10,0.17260250449180603,1.2378572225570679,0.59874629974365234
-11,0.25938931107521057,1.4522590637207031,0.7777174711227417
-12,0.35764923691749573,1.5975272655487061,0.90651595592498779
-13,0.47831591963768005,1.721396803855896,1.0185583829879761
-14,0.62966132164001465,1.8401021957397461,1.1312437057495117
-15,0.81133067607879639,1.952968955039978,1.2502502202987671
-16,1.0182336568832397,2.0505752563476562,1.3766647577285767
-17,1.2369400262832642,2.1172373294830322,1.5072586536407471
-18,1.4501547813415527,2.1338057518005371,1.6361281871795654
-19,1.6380794048309326,2.0827107429504395,1.7541790008544922
-20,1.7830932140350342,1.9511455297470093,1.8499293327331543
-21,1.8726725578308105,1.7365177869796753,1.9103424549102783
-22,1.9004602432250977,1.4457336664199829,1.9218668937683105
-23,1.8667293787002563,1.0954406261444092,1.873089075088501
-24,1.7771525382995605,0.70876187086105347,1.7590262889862061
-25,1.6419719457626343,0.30778434872627258,1.5867030620574951
-26,1.4697904586791992,-0.11207789182662964,1.3660774230957031
-27,1.256860613822937,-0.62711113691329956,1.0884270668029785
-28,0.96113878488540649,-1.5642948150634766,0.67682826519012451
-29,0.38500988483428955,-6.0011897087097168,-0.29292356967926025
-30,-2.5814776420593262,5.9809556007385254,-29.186309814453125
-31,4.6067681312561035,2.3116941452026367,3.520749568939209
-32,2.6724138259887695,1.5919616222381592,2.456402063369751
-33,2.2848775386810303,1.4409234523773193,2.1952269077301025
-34,2.2064304351806641,2.2376229763031006,2.258394718170166
-35,2.3257617950439453,-0.84194576740264893,2.8443450927734375
-36,2.7335648536682129,0.16693255305290222,8.7644834518432617
-37,4.1261763572692871,0.40835437178611755,-1.4980161190032959
-38,-20.82036018371582,0.58964473009109497,0.11774235963821411
-39,-0.27205550670623779,0.78705590963363647,0.62460339069366455
-40,0.87708234786987305,1.0215424299240112,0.94127070903778076
-41,1.3943554162979126,1.2936074733734131,1.205518364906311
-42,1.7287877798080444,1.5936223268508911,1.4548220634460449
-43,1.9613333940505981,1.9005235433578491,1.6968636512756348
-44,2.1115937232971191,2.1884396076202393,1.9276688098907471
-45,2.1847262382507324,2.4287583827972412,2.135878324508667
-46,2.1843020915985107,2.5953071117401123,2.3066277503967285
-47,2.1161854267120361,2.6678018569946289,2.4239785671234131
-48,1.9869258403778076,2.63327956199646,2.4731285572052002
-49,1.8022885322570801,2.4894053936004639,2.4433183670043945
